BSE: Make gslwaveosc work for 96k samples (fixes bug #612281).
For a 96k sample, using a filter transition from 18kHz to 24kHz is not optimal
because this transition band is quite steep, and thus the stopband ripple will
be quite high. However, for a 96kHz sample we can set the start of the stopband
to 48kHz, which results in a less steep filter, and better stopband rejection.
This change improves stopband settings for any wave file with a sample rate >
48 hHz.