BSE: implemented new concept: BseStorageBlob
A blob is a binary large object, that is some binary data which is stored along
with the other data. It was designed to store sound font files within .bse
files, while still allowing fluid synth to access the file directly. Temp files
are used to give fluid synth direct access to a file.