BSE: fix bseloader-bsewave.cc for g++-4.4
For some reason, g++-4.4 will not convert the token value of an enum to the
"right" number if the enum is converted to an uint; it does work for enum->int,
though. So this change will make the bsewave parser work for g++-4.4; I'm not
sure whether this is a compiler bug, or whether the C++ language specification
allows the behaviour g++-4.4 exposes (the old code before this commit passes
the testsuite when compiled with g++-4.5, but breaks it for g++-4.4).