TRAVIS: redo snapshot versioning to follow debian guidelines more closely
The monotonic git revision becomes part of the package upstream version as
it indicates upstream patches. It is added so it's sorted *before* the next
devel release or *after* the last stable. The distribution version is now
part of the package build revision, which reflects the different builds we
are starting.

Signed-off-by: Tim Janik <timj@gnu.org>