BSE: fix C++ property loading (float64)
The old (buggy) code called scanner_parse_paren_rest, and then an string to
double conversion. It was assumed that scanner_parse_paren_rest would just
return the tokens concatenated in one result string. Here is what happens:

Input: 1.03

1) scanner sees integer token, rest is now "1"
2) scanner sees dot, rest is now "1."
3) scanner sees integer token
  3a) scanner appends space to rest, because rest in non-empty, rest is now "1. "
  3b) scanner appends value of the integer, rest is now "1. 3"

So the buggy code parses this as 1.0 (not 1.03).
Even if (3a) wouldn't append " ", this would parse as 1.3 (not 1.03).

Not sure if scanner_parse_paren_rest could be modified, what I did is to
special case float64 values, and use the standard sfi way of parsing this. I
tested that this works for all C++ float64 bus properties tested in make check.

Signed-off-by: Stefan Westerfeld <stefan@space.twc.de>