BSE: bseserver: add shared_block_offset() and increase shm area preparations
Using shared_block_offset(), the offset of a memory area internal pointer
can be reconstructed which can simplify eternal book keeping. Also, this
increases shm area granularity to cache-line size, so we don't waste more
than a cache-line for smaller allocations like we need for DSP parameters.
Preallocate the shm area as 8MB (that corresponds to 128k cache lines),
because it'll need to contain all possible interface port allocations
throughout the runtime and might not be resizable on some systems.

Signed-off-by: Tim Janik <timj@gnu.org>