EBEAST: b/devicepanel.vue: make device area scrollable with nice scrollbar
Making the device panel scrollable and look nice without an oddly offset
scrollbar is not easy. So instead make the scrollbar slim and dim it if
the panel is not in hover state. Scrollbar styling is tricky however,
Chrome supports a pixel widths and setting gradients and borders for
various scrollbar parts. Firefox just allows a "thin" scrollbar toggle
and plain colors, but supports color fading on hover.

Signed-off-by: Tim Janik <timj@gnu.org>