DEVICES: freeverb: implement STK and VLC damping
Avoid reinforcing feedback loop in the original damping code,
by providing VLC and STK damping modes. The latter is likely
closest to the original intention so it is now the default.
This also reorders some parameters and removes the original
freezing mode, that does not really make sense for Processors.

Signed-off-by: Tim Janik <timj@gnu.org>