DEVICES: blepsynth/blepsynth.cc: add oscillator shape plotting code
The code from plotblep.py was C++ified, in order to be able to display
Osc1/2 waveform at the UI. As there is no UI support for that yet, a
gnuplot file will be written. The code needs to be explicitely enabled,
as it dumps the files in the RT processor code.

Signed-off-by: Stefan Westerfeld <stefan@space.twc.de>